refactor(geometry): split processors/tessellated.rs into 5 concern modules
tessellated.rs (831 LOC) -> tessellated/{mod, triangulated, polygonal,
triangulate (self-contained earcut), mesh_build (shared flat-shaded mesh
builders)}. All files under ~400. Pure code motion; triangulate_polygon ->
pub(super), extract_coord_index_bytes reached via super::super (no mod.rs
change). processors/mod.rs + siblings untouched. build/clippy clean; tests green.
